Moved model classes to file package
