+++ web/src/main/java/com/epam/rft/atsy/web/model/file/FileStatus.java
@@ -1,0 +1,15 @@
+package com.epam.rft.atsy.web.model.file;
+
+import lombok.AllArgsConstructor;
+
+@AllArgsConstructor
+public enum FileStatus {
+
+  FILE_IS_NOT_EXIST(false),
+  FILE_IS_IN_PROGRESS(null),
+  FILE_IS_ALREADY_EXIST(true);
+
+  private Boolean value;
+
+  public Boolean getValue() { return value; }
+}
